Move Mock database location
