+++ app/src/androidTest/java/com/github/se/studybuddies/tests/AccountSettingsTest.kt
@@ -5,7 +5,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.se.studybuddies.navigation.NavigationActions
 import com.github.se.studybuddies.navigation.Route
 import com.github.se.studybuddies.screens.AccountSettingsScreen
-import com.github.se.studybuddies.testUtilities.fakeDatabase.MockDatabase
+import com.github.se.studybuddies.database.MockDatabase
 import com.github.se.studybuddies.ui.account.AccountSettings
 import com.github.se.studybuddies.viewModels.UserViewModel
 import com.kaspersky.components.composesupport.config.withComposeSupport
