+++ app/src/androidTest/java/com/github/se/studybuddies/tests/CreateAccountTest.kt
@@ -6,7 +6,7 @@ import androidx.test.espresso.Espresso.closeSoftKeyboard
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.se.studybuddies.navigation.NavigationActions
 import com.github.se.studybuddies.navigation.Route
-import com.github.se.studybuddies.testUtilities.fakeDatabase.MockDatabase
+import com.github.se.studybuddies.database.MockDatabase
 import com.github.se.studybuddies.ui.account.CreateAccount
 import com.github.se.studybuddies.viewModels.UserViewModel
 import com.kaspersky.components.composesupport.config.withComposeSupport
