+++ app/src/androidTest/java/com/github/se/studybuddies/tests/LeaveGroupTest.kt
@@ -3,7 +3,7 @@ package com.github.se.studybuddies.tests
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.se.studybuddies.navigation.NavigationActions
-import com.github.se.studybuddies.testUtilities.fakeDatabase.MockDatabase
+import com.github.se.studybuddies.database.MockDatabase
 import com.github.se.studybuddies.ui.groups.GroupsSettingsButton
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
 import io.mockk.impl.annotations.RelaxedMockK
