+++ app/src/androidTest/java/com/github/se/studybuddies/tests/MapTest.kt
@@ -9,7 +9,7 @@ import com.github.se.studybuddies.data.User
 import com.github.se.studybuddies.database.DatabaseConnection
 import com.github.se.studybuddies.navigation.NavigationActions
 import com.github.se.studybuddies.screens.MapScreen
-import com.github.se.studybuddies.testUtilities.fakeDatabase.MockDatabase
+import com.github.se.studybuddies.database.MockDatabase
 import com.github.se.studybuddies.ui.map.MapScreen
 import com.github.se.studybuddies.viewModels.UserViewModel
 import com.github.se.studybuddies.viewModels.UsersViewModel
