+++ app/src/main/java/com/github/se/studybuddies/testUtilities/MockMainActivity.kt
@@ -2,13 +2,21 @@ package com.github.se.studybuddies.testUtilities
 
 import android.os.Bundle
 import com.github.se.studybuddies.MainActivity
-import com.github.se.studybuddies.testUtilities.fakeDatabase.MockDatabase
+import com.github.se.studybuddies.database.DatabaseConnection
+import com.github.se.studybuddies.database.MockDatabase
+import com.google.firebase.auth.FirebaseAuth
 
 class MockMainActivity : MainActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
-    var fakeUID = "E2EUserTest"
+    val fakeUID = "E2EUserTest"
     db = MockDatabase()
     startApp(fakeUID, db)
+  }
+  override fun onStop() {
+    super.onStop()
+    val fakeUID = "E2EUserTest"
+    db = MockDatabase()
+    offlineLocation(fakeUID, db)
   }
 }
